[DS-3909] Patch support for withdraw, reinstate and discoverable added to Item repository.
[DS-3909] Updated the patch method and helper methods.

The patch method response returns json for the updated item.  Not sure this is needed.

[DS-3909] Integration tests.

The tests inspect the full item json as returned by the current patch implementation.

[DS-3909] Removed add eperson to context from StatelessAuthenticationFilter.

This addition was suggested by @bollini as a fix for a problem I was having earlier. I verified that my code now works without this addition, so I propose removing it.