migrate empty nested property values
generate placeholder values for nested properties, that does not exist
within a single nested object. The Placeholder are generated by
considering the tp2pdef table in some weird join query to create values
for the properties that can possible be assigned but are not assigned
yet. With these placeholders migrated metadatagroups are displayed
correctly.